third_party/WebKit/LayoutTests/compositing/copy-image-when-no-image-exists-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS width is 0
PASS height is 0
